[CL-1941] Rename ContentBuilderPreview to ContentViewer
