+++ front/app/modules/commercial/content_builder/admin/components/ContentViewer/Viewer/index.tsx
